pyproject.toml
[tool.poetry]
name = "enterprise-qa-assistant"
version = "0.1.0"
description = "Enterprise QA Assistant with RAG and LangGraph orchestration"
authors = ["Your Name <your@email.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# === Web Framework ===
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sse-starlette = "^2.1.0"

# === LangChain / LangGraph / LlamaIndex ===
langchain = "^1.2.0"
langchain-openai = "^1.1.2"
langchain-community = "^0.3.0"
langgraph = "^1.1.0"
langgraph-checkpoint = "^4.0.0"
llama-index = "^0.14.0"
llama-index-llms-openai = "^0.7.0"
llama-index-readers-file = "^0.5.0"

# === Database ===
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.0"
pymilvus = "^2.4.0"
elasticsearch = "8.12.0"

# === Infrastructure ===
structlog = "^24.0.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
arq = "^0.25.0"

# === Document Processing ===
pypdf = "^5.0.0"
python-docx = "^1.1.0"
markdown = "^3.7"

# === Utilities ===
python-multipart = "^0.0.9"
aiofiles = "^24.0.0"
tenacity = "^9.0.0"

# === Authentication ===
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
bcrypt = "^4.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
ruff = "^0.7.0"
mypy = "^1.13.0"
pre-commit = "^4.0.0"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
